Awards that are dependent on bronze/silver Badges have been fixed
Awards that needed a specific version of a Badge (Bronze or Silver) unlocked could not find them in the database, if the higher version of that Badge has already been rewarded. badge/Medal checks now not only consider the bronze Badge, but the silver and gold one aswell.
Fixed many typos and co-op related stuff too.